Implemented unlimited Braid-style cloning
Also removed the need for a recording manager. The current Event system
holds a target positionToMoveTo variable that will be passed to any
clones. If the player rewinds and creates a clone before he reaches his
target destination, the spawned clone will follow through with that
Event.
This leads to an interesting feature where the player can quickly click
where he wants to go, immediately spawn a clone, and have the clone
follow through with that action.

TODOs left:
- Implement rotation
- Implement interactivity on the Event class.
- Destroy clones when a limit is reached